+++ pkg/apis/clouddns/v1/xxgenerated_object.go
@@ -11,6 +11,11 @@ func (o *Record) GetIdentifier(ctx context.Context) (string, error) {
 	return o.Identifier, nil
 }
 
+// GetIdentifier returns the primary identifier of a Revision object
+func (o *Revision) GetIdentifier(ctx context.Context) (string, error) {
+	return o.Identifier, nil
+}
+
 // GetIdentifier returns the primary identifier of a Zone object
 func (o *Zone) GetIdentifier(ctx context.Context) (string, error) {
 	return o.Name, nil